Implement optional light resource mode
When the --light cli arg is specified, the indexer doesn't store transactions,
block meta data (tx count, size and weight) and the blockhash=>txids map.

Fallbacks that use the bitcoind rpc to fetch information when the extended
indexes are unavailable was implemented in Query for load_txn, get_block_meta
and get_block_txids.

This saves up around 250GB of storage, at the cost of more expensive lookups
and more reliance on bitcoind.

To reduce memory and CPU costs, the --disable-prevout parameter can be
specified to disable attaching previous output information to inputs.
This significantly reduces the amount of transaction lookups, at the
cost of missing inputs amounts/addresses and transaction fees.

This commit also introduces db versioning and a compatibility field that
requires users to reindex the db after switching networks or changing light mode.