Treat issuance amount of null as zero
Fixes https://github.com/Blockstream/esplora/issues/146

It appears like issuances of 0 asset units (like in [0] from #146) are encoded
as a null value rather than as an explicit value of zero, so we convert it to
a zero ourselves. This matches the behaviour of the re-issuance tokens, which
also represent zero tokens as a null value (which we already convert to zero).

[0] https://blockstream.info/liquid/tx/152a6bce79d49acfea3fc9f3c905f3580a646cbadea2f07fec93d9462d7d2731/