Set Rust tool chain to 1.41.1
The project does not compile on 1.34.0 because it uses the lockfile
format introduced in 1.38.0.
There are also issues encountered with versions 1.39.0 to 1.43.0 as
seen here: D4nte/electrs/runs/1174430513

To stay compatible with Debian stable (10 at the time), we want
this project to run with 1.41.1.

Setting a specific toolchain version is good practice for binary
projects and remove potential issues of incompatible toolchain for
users trying to build the project.