Use sync primitives from parking_lot
This change adds a custom module mimicking `std::sync` which may contain
custom implementations of synchronization primitives. It also redirects
all cases of `std::sync` to use this new module and replaces mutexes
with thosse from `parking_lot`.

The main motivation is to help with finding deadlocks using deadlock
detection feature from `parking_lot`. The locks from `parking_lot` don't
implement poisoning so we needed to add it on top to maintain perfect
compatibility. This was solved by wrapping locks in custom types which
mirror the `std` API.

Deadlock detection is behind a feature flag to not slow donw production
use and also because it's considered experimental.

Because this change makes swapping primitives easy, we could try
replacing more of them in the future and measure performance impact.
However I don't expect significant changes and we should focus on
low-hanging fruit first.