Fix format checking logic for legacy databases
Old versions of `electrs` didn't contain format information and
previously the code would use default which is the current one thus
mistakenly assume the version is recent. This would lead to blow up of
the database.

This fixes the problem by checking if the database is empty and assuming
it's legacy if it's not empty but without format information.