Detailed documentation of db_dir
This warns users that they shouldn't touch `db_dir` internals except for
deleting the old `mainnet` database. It explains in details how to do
upgrade and deleting properly, calls it out in release notes and fixes
small style mistake in nearby heading.

It can be thought of as more explicit/detailed version of
8b8ebea0018d3e9a97ce965845b1914df4b4d8a4 or a textual alternative to
PR #513.