Improve config documentation
The `--skip-default-conf-files` flag is often desired but people may miss it because it's not in the doc, so this documents it.

Further, to avoid problems with people missing flags this adds a recommendation to run `electrs --help` so if we ever forget to update the doc people will learn about the new arguments anyway.

Following #962.