logically separate mempool sync polling and mutation
The existing Mempool::sync logic maintained a mutable reference
to the Mempool even though it's not actually using that
mutable privilege for the majority of the time that the sync
method is running. We separate the fetching the polling and
applying processes, so that we can make mempool polling async.