offer alternative methods for customizing TX broadcast privacy
Transactions which are first seen being broadcast from a user's node are
more likely to be heuristically associated with the node runner personally.
To break this heuristic, node runners might choose to broadcast a TX
indirectly through alternative channels to disassociate that TX from
the bitcoin node.

This PR adds two new ways for users to broadcast transactions through electrs:

1. The `pushtx` crate. This initiates a short-lived set of P2P connections
which publishes a transaction to random peers resolved from DNS seeds. The
most private way to do this is using TOR, hiding the IP address of the sender.

2. A custom script option. This allows users to customize the behavior of
electrs broadcasting. It should return a zero status if the broadcast was
successful, and emit an error message with a non-zero status otherwise.