Fix Elements dynafed tests
They were broken by the test added in the last commit, which mined an
extra block and caused the `GET /block/:hash` tests for Elements to run
against a Compact params blocks instead of a Full one.

This commit fixes it by running the tests against blocks 1 & 2 which
are known to be Full & Compact respectively.